Refactor Flask routes into Blueprints
